Resolve conflicts, refactor deployAssimilators and Utils files, and add ConfigValidator file
