+++ scripts/ConfigValidator.ts
@@ -13,3 +13,35 @@ export const validateAssimilatorConfig = config => {
 
   return true;
 }
+
+export const validateCurveConfig = params => {
+  const dimensionKeys = [
+    'alpha',
+    'beta',
+    'max',
+    'epsilon',
+    'lambda'
+  ];
+
+  const paramsKeys = [
+    'token_symbol',
+    'token_name',
+    'weights',
+    'lpt-name',
+    'dimensions'
+  ];
+
+  for (const paramKey of paramsKeys) {
+    const extractedParamsKeys = Object.keys(params);
+    if (!extractedParamsKeys.includes(paramKey)) return false;
+
+    if (extractedParamsKeys.includes('dimensions') && paramKey === 'dimensions') {
+      const extractedDimensionKeys = Object.keys(params.dimensions);
+      for (const dimensionKey of dimensionKeys) {
+        if (!extractedDimensionKeys.includes(dimensionKey)) return false;
+      }
+    }
+  }
+
+  return true;
+}
